Fix [2586695]: you can no longer throw items (including via Portal Projectile) if you can't wield weapons. Don't allow base-shape breath attacks when shapechanged, except for Nagas in spider form. Fix [2589913]: melded shields counted for heavy armour penalties.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@9026 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/describe.cc
@@ -595,7 +595,7 @@ const char *trap_name(trap_type trap)
     ASSERT(NUM_TRAPS == sizeof(trap_names) / sizeof(*trap_names));
 
     if (trap >= TRAP_DART && trap < NUM_TRAPS)
-        return trap_names[ static_cast<int>( trap ) ];
+        return trap_names[trap];
     return (NULL);
 }
